requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain-openai
python-dotenv
huggingface_hub
# langchain_huggingface
langchain-huggingface